# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_jhoicas_Inventario-api_internal_application_dto.CreateCustomerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      tax_id:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.CustomerListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CustomerResponse'
        type: array
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.CustomerResponse:
    properties:
      category_name:
        type: string
      company_id:
        type: string
      email:
        type: string
      id:
        type: string
      ltv:
        type: number
      name:
        type: string
      phone:
        type: string
      tax_id:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
...
  github_com_jhoicas_Inventario-api_internal_application_dto.UpdateCustomerRequest:
    properties:
      email:
        type: string
      name:
        type: string
      phone:
        type: string
      tax_id:
        type: string
    type: object
...
host: api.naturerp.ludoia.com
info:
  contact: {}
  description: API para el sistema ERP.
  title: Tu API ERP
  version: "1.0"
paths:
...
  /api/customers:
    get:
      consumes:
      - application/json
      description: Lista los clientes de la empresa autenticada con paginación
      parameters:
      - description: Buscar por nombre o NIT (tax_id)
        in: query
        name: search
        type: string
      - description: Límite de resultados
        in: query
        name: limit
        type: integer
      - description: Desplazamiento
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CustomerListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Listar clientes
      tags:
      - customers
    post:
      consumes:
      - application/json
      description: Crea un cliente asociado a la empresa autenticada para uso en facturación
      parameters:
      - description: Datos del cliente
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CreateCustomerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Crear cliente
      tags:
      - customers
  /api/customers/{id}:
    put:
      consumes:
      - application/json
      description: Actualiza un cliente asociado a la empresa autenticada
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Datos del cliente
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.UpdateCustomerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.CustomerResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Actualizar cliente
      tags:
      - customers
  /api/customers/{id}/deactivate:
    put:
      description: Desactiva (soft delete) un cliente asociado a la empresa autenticada
        (solo admin)
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_jhoicas_Inventario-api_internal_application_dto.ErrorResponse'
      security:
      - Bearer: []
      summary: Desactivar cliente
      tags:
      - customers
...
securityDefinitions:
  Bearer:
    description: |-
      Escribe "Bearer " seguido de un espacio y tu token.
      Escribe "Bearer " seguido de un espacio y tu token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"